requirements.txt
dj-database-url==0.4.0
Django==1.9.2
django-filter==0.12.0
djangorestframework==3.3.2
gunicorn==19.4.5
<<<<<<< Updated upstream
psycopg2==2.6.1
sendgrid==1.6.20
=======
wheel==0.26.0
>>>>>>> Stashed changes
whitenoise==2.0.6
